Support for packing multiple inputs into one output
This is the first pass at packing. `chain_pack(objects, buffer)` will chain
transforms on all objects in `objects` so that they are in a straight line
in the center aligned on the X axis.

The transformations are not automatically applied.

The `-p` command line option enables packing
The `-b <number>` changes the default padding of 5.0